perf: 64-bit class bloom filter, XPath lazy iterator
4D.3: Expand per-node class_hash from u32 to u64
  - Reduces false positive rate from ~15% (5 classes) to ~8%
  - class_bloom_bit() now returns u64, hash % 64
  - Node struct: class_hash u32→u64, padding 11→7 bytes
  - All consumers updated (arena, matcher, AST, parser)

4C.4: XPath absolute path evaluation — eliminate per-step Vec allocations
  - Reuse two buffers (current + next) with swap instead of allocating per step
  - collect_element_children() writes to caller's buffer instead of returning Vec
  - Remove unused direct_element_children wrapper

+++ crates/fhp-selector/src/ast.rs
@@ -42,8 +42,8 @@ pub enum SimpleSelector {
     Tag(Tag),
     /// Match by unknown/custom tag name, e.g. `my-widget`.
     UnknownTag(String),
-    /// Match by class: `.class`. Second field is precomputed bloom bit.
-    Class(String, u32),
+    /// Match by class: `.class`. Second field is precomputed 64-bit bloom bit.
+    Class(String, u64),
     /// Match by id: `#id`. Second field is precomputed FNV-1a hash.
     Id(String, u32),
     /// Universal selector: `*`.
